backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database - Snowflake
snowflake-sqlalchemy>=1.6.1  # 1.6.0 not available, using 1.6.1+
snowflake-connector-python==3.7.0
sqlalchemy==2.0.23
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Google APIs
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
google-generativeai>=0.3.0  # Gemini API (optional - we use httpx, but this is available)

# AI/ML - Dedalus Labs & K2-Think
# dedalus-labs==0.1.0  # Not available on PyPI yet - install manually when available
requests>=2.31.0
httpx>=0.25.1

# Data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# CORS
# python-cors==1.0.0  # Not needed - FastAPI has built-in CORS support

# Testing (optional for now)
pytest==7.4.3
pytest-asyncio==0.21.1
